+++ src/of_sexp.ml
@@ -109,14 +109,14 @@ let date_time_of_sexp (x : CCSexp.t) =
           | None -> invalid_data ())
       | `Tz_offset_s_only tz_offset_s -> (
           match
-            Time.Date_time'.make_precise ~year ~month ~day ~hour ~minute ~second
+            Time.Date_time'.make_unambiguous ~year ~month ~day ~hour ~minute ~second
               ~ns ~tz_offset_s ()
           with
           | Some x -> x
           | None -> invalid_data ())
       | `Tz_and_tz_offset_s (tz, tz_offset_s) -> (
           match
-            Time.Date_time'.make_precise ~tz ~year ~month ~day ~hour ~minute
+            Time.Date_time'.make_unambiguous ~tz ~year ~month ~day ~hour ~minute
               ~second ~ns ~tz_offset_s ()
           with
           | Some x -> x